requirements.txt
asgiref==3.5.0
Django==4.0.2
numpy==1.22.2
Pillow==9.0.1
pytesseract==0.3.8
sqlparse==0.4.2
tzdata==2021.5